+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A `mix()` iterator.
 - A CAS-register generator.
 - A bank generator.
+- A `cycle_times()` iterator.
 
 ### Changed